Handle everything via golang
