+++ README.md
@@ -1,8 +1,8 @@
 # crd-codegen
 
-**Work in progress**
+Generates Go type definitions for a Kubernetes Custom Resource by parsing a YAML `CustomResourceDefinition`. Useful if an operator/application was not written in Go but you have a need to interact with its resoures.
 
-Generates Go API interfaces for a Kubernetes Custom Resource by parsing a YAML `CustomResourceDefinition`. Useful if an operator/application was not written in Go but you have a need to interact with its resoures.
+This tool extracts the OpenAPIv3 spec from the CRD and runs it through a JSON schema type generator.
 
 ## Example usage
 
@@ -10,7 +10,7 @@ Using the Strimzi operator as an example:
 
 ```
 # grab dependencies
-go get
+go get -u github.com/bsquizz/crd-codegen/...
 export PATH=$PATH:$(go env GOPATH)
 
 # download CRD's for strimzi operator
@@ -18,10 +18,8 @@ VERSION=0.21.1
 wget https://github.com/strimzi/strimzi-kafka-operator/releases/download/${VERSION}/strimzi-crds-${VERSION}.yaml
 
 # generate types using the CRD
-./generate.sh strimzi-crds-${VERSION}.yaml
+crd-codegen -in=strimzi-crds-${VERSION}.yaml -out=generated
 
-# generated apis are found in 'generated' directory:
 ls generated/apis/kafka.strimzi.io/v1beta1/
 
-# KafkaConnect.go  KafkaConnectS2I.go  Kafka.go  KafkaMirrorMaker.go  KafkaTopic.go  KafkaUser.go
 ```
