+++ go.mod
@@ -1,10 +1,12 @@
-module github.com/bsquizz/strimzi-go
+module github.com/bsquizz/crd-codegen
 
 go 1.15
 
 require (
-	github.com/atombender/go-jsonschema v0.8.0 // indirect
-	github.com/mikefarah/yq/v4 v4.4.1 // indirect
+	github.com/a-h/generate v0.0.0-20190312091541-e59c34d33fb3 // indirect
+	github.com/orus-io/json-schema-generate v0.0.0-20200422145733-cc402b93ccf6
+	github.com/shopspring/decimal v1.2.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0
 	k8s.io/apiextensions-apiserver v0.20.2
 	k8s.io/apimachinery v0.20.2
 	k8s.io/client-go v0.20.2
